requirements.txt
asgiref==3.2.10
boto3==1.14.50
botocore==1.17.50
Cerberus==1.3.2
dj-database-url==0.5.0
Django==3.1
django-cors-headers==3.5.0
django-filter==2.3.0
django-heroku==0.3.1
django-storages==1.9.1
djangorestframework==3.11.1
docutils==0.15.2
Faker==4.1.2
gunicorn==20.0.4
jmespath==0.10.0
model-bakery==1.1.1
Pillow
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.10
whitenoise==5.2.0
# scikit-build
# cmake
# pytesseract
# opencv-contrib-python
